// src/common/events/constants/socket-error-codes.enum.ts
export enum SocketErrorCode {
  ROOM_NOT_FOUND = 'ROOM_NOT_FOUND',
  NOT_ROOM_OWNER = 'NOT_ROOM_OWNER',
  NOT_MODERATOR = 'NOT_MODERATOR',
  ALREADY_IN_QUEUE = 'ALREADY_IN_QUEUE',
  NOT_IN_QUEUE = 'NOT_IN_QUEUE',
  INVALID_EMOJI = 'INVALID_EMOJI',
  USER_NOT_ON_STAGE = 'USER_NOT_ON_STAGE',
  ALREADY_ON_STAGE = 'ALREADY_ON_STAGE',
  ROOM_CLOSED = 'ROOM_CLOSED',
  UNAUTHORIZED = 'UNAUTHORIZED',
  NO_INVITATION_FOUND = 'NO_INVITATION_FOUND',
  INVALID_ROOM_STATE = 'INVALID_ROOM_STATE',
  NOT_IN_ROOM = 'NOT_IN_ROOM',
  ROOM_BANNED = 'ROOM_BANNED',
  ROOM_FULL = 'ROOM_FULL',
  ROOM_LOCKED = 'ROOM_LOCKED',
  INVITATION_REQUIRED = 'INVITATION_REQUIRED',
  TICKET_REQUIRED = 'TICKET_REQUIRED',
  SUBSCRIPTION_REQUIRED = 'SUBSCRIPTION_REQUIRED',
  VERIFICATION_REQUIRED = 'VERIFICATION_REQUIRED',
  CLUB_MEMBERSHIP_REQUIRED = 'CLUB_MEMBERSHIP_REQUIRED',
}